+++ pkg/xds/envoy/listeners/http_access_log_configurer_test.go
@@ -32,6 +32,7 @@ var _ = Describe("HttpAccessLogConfigurer", func() {
 			mesh := "demo"
 			sourceService := "web"
 			destinationService := "backend"
+			trafficDirection := "OUTBOUND"
 			proxy := &core_xds.Proxy{
 				Id: xds.ProxyId{
 					Name: "web",
@@ -61,7 +62,7 @@ var _ = Describe("HttpAccessLogConfigurer", func() {
 				Configure(FilterChain(NewFilterChainBuilder().
 					Configure(HttpConnectionManager(given.statsName)).
 					Configure(HttpOutboundRoute(given.routeName)).
-					Configure(HttpAccessLog(mesh, sourceService, destinationService, given.backend, proxy)))).
+					Configure(HttpAccessLog(mesh, trafficDirection, sourceService, destinationService, given.backend, proxy)))).
 				Build()
 			// then
 			Expect(err).ToNot(HaveOccurred())