+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## master
+* feat include traffic direction in access log
+  [#682](https://github.com/Kong/kuma/pull/682)
 * feat: validate tags and selectors
   [#691](https://github.com/Kong/kuma/pull/691) 
 * feat: refactor CA to plugins